Allow to skip a number of launch.sh stages (data localization, pipe installation, nfs installation)

+++ scripts/pipeline-launch/launch.sh
@@ -1062,32 +1062,35 @@ if [ -d $COMMON_REPO_DIR/shell ]; then
 fi
 
 # Install pipe CLI
-if [ "$CP_PIPELINE_CLI_FROM_DIST_TAR" ]; then
-      install_pip_package PipelineCLI
-else
-      echo "Installing 'pipe' CLI"
-      echo "-"
-      CP_PIPELINE_CLI_BINARY_NAME="${CP_PIPELINE_CLI_BINARY_NAME:-pipe}"
-      download_file "${DISTRIBUTION_URL}${CP_PIPELINE_CLI_BINARY_NAME}"
-      if [ $? -ne 0 ]; then
-            echo "[ERROR] 'pipe' CLI download failed. Exiting"
-            exit 1
+CP_PIPE_CLI_ENABLED=${CP_PIPE_CLI_ENABLED:-"true"}
+if [ "$CP_PIPE_CLI_ENABLED" == "true" ]; then
+      if [ "$CP_PIPELINE_CLI_FROM_DIST_TAR" ]; then
+            install_pip_package PipelineCLI
+      else
+            echo "Installing 'pipe' CLI"
+            echo "-"
+            CP_PIPELINE_CLI_BINARY_NAME="${CP_PIPELINE_CLI_BINARY_NAME:-pipe}"
+            download_file "${DISTRIBUTION_URL}${CP_PIPELINE_CLI_BINARY_NAME}"
+            if [ $? -ne 0 ]; then
+                  echo "[ERROR] 'pipe' CLI download failed. Exiting"
+                  exit 1
+            fi
+
+            # Clean any known locations, where previous verssion of the pipe might reside (E.g. committed by the user)
+            rm -f /bin/pipe
+            rm -f /usr/bin/pipe
+            rm -f /usr/local/bin/pipe
+            rm -f /sbin/pipe
+            rm -f /usr/sbin/pipe
+            rm -f /usr/local/sbin/pipe
+            rm -f ${CP_USR_BIN}/pipe
+
+            # Install into the PATH locationse
+            cp pipe /usr/bin/
+            cp pipe ${CP_USR_BIN}/
+            chmod +x /usr/bin/pipe ${CP_USR_BIN}/pipe
+            rm -f pipe
       fi
-
-      # Clean any known locations, where previous verssion of the pipe might reside (E.g. committed by the user)
-      rm -f /bin/pipe
-      rm -f /usr/bin/pipe
-      rm -f /usr/local/bin/pipe
-      rm -f /sbin/pipe
-      rm -f /usr/sbin/pipe
-      rm -f /usr/local/sbin/pipe
-      rm -f ${CP_USR_BIN}/pipe
-
-      # Install into the PATH locationse
-      cp pipe /usr/bin/
-      cp pipe ${CP_USR_BIN}/
-      chmod +x /usr/bin/pipe ${CP_USR_BIN}/pipe
-      rm -f pipe
 fi
 
 # Install FS Browser
@@ -1257,28 +1260,28 @@ echo
 ######################################################
 
 
-if [ "$RESUMED_RUN" == true ];
-then
-    echo "Skipping data localization for resumed run"
-else
-    ######################################################
-    echo "Checking if remote data needs localizing"
-    echo "-"
-    ######################################################
-    LOCALIZATION_TASK_NAME="InputData"
-    INPUT_ENV_FILE=${RUN_DIR}/input-env.txt
+CP_DATA_LOCALIZATION_ENABLED=${CP_DATA_LOCALIZATION_ENABLED:-"true"}
+if [ "$CP_DATA_LOCALIZATION_ENABLED" == "true" ]; then
+      if [ "$RESUMED_RUN" == true ]; then
+            echo "Skipping data localization for resumed run"
+      else
+            ######################################################
+            echo "Checking if remote data needs localizing"
+            echo "-"
+            ######################################################
+            LOCALIZATION_TASK_NAME="InputData"
+            INPUT_ENV_FILE=${RUN_DIR}/input-env.txt
 
-    upload_inputs "${INPUT_ENV_FILE}" "${LOCALIZATION_TASK_NAME}"
+            upload_inputs "${INPUT_ENV_FILE}" "${LOCALIZATION_TASK_NAME}"
 
-    if [ $? -ne 0 ];
-    then
-        echo "Failed to upload input data"
-        exit 1
-    fi
-    echo
+            if [ $? -ne 0 ]; then
+                  echo "Failed to upload input data"
+                  exit 1
+            fi
+            echo
 
-    [ -f "${INPUT_ENV_FILE}" ] && source "${INPUT_ENV_FILE}"
-
+            [ -f "${INPUT_ENV_FILE}" ] && source "${INPUT_ENV_FILE}"
+      fi
 fi
 echo "------"
 echo
